finished rework round
well... there's one bug I noticed I want to go ahead and fix.
